+++ bakerydemo/locations/models/__init__.py
@@ -1,0 +1,3 @@
+from .operating_hours import *  # noqa
+from .location_page import *  # noqa
+from .location_index_page import *  # noqa
