+++ bakerydemo/locations/models/location_index_page.py
@@ -1,0 +1,47 @@
+from django.db import models
+from wagtail.admin.panels import FieldPanel
+from wagtail.images.edit_handlers import ImageChooserPanel
+from wagtail.models import Page
+
+from bakerydemo.locations.models import LocationPage
+
+
+class LocationsIndexPage(Page):
+    """
+    A Page model that creates an index page (a listview)
+    """
+    introduction = models.TextField(
+        help_text='Text to describe the page',
+        blank=True)
+    image = models.ForeignKey(
+        'wagtailimages.Image',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+        help_text='Landscape mode only; horizontal width between 1000px and 3000px.'
+    )
+
+    # Only LocationPage objects can be added underneath this index page
+    subpage_types = ['LocationPage']
+
+    # Allows children of this indexpage to be accessible via the indexpage
+    # object on templates. We use this on the homepage to show featured
+    # sections of the site and their child pages
+    def children(self):
+        return self.get_children().specific().live()
+
+    # Overrides the context to list all child
+    # items, that are live, by the date that they were published
+    # https://docs.wagtail.org/en/stable/getting_started/tutorial.html#overriding-context
+    def get_context(self, request):
+        context = super(LocationsIndexPage, self).get_context(request)
+        context['locations'] = LocationPage.objects.descendant_of(
+            self).live().order_by(
+            'title')
+        return context
+
+    content_panels = Page.content_panels + [
+        FieldPanel('introduction', classname="full"),
+        ImageChooserPanel('image'),
+    ]
